+++ app/src/main/java/com/godamy/myplayer/ui/detail/DetailInfoView.kt
@@ -6,7 +6,7 @@ import androidx.appcompat.widget.AppCompatTextView
 import androidx.core.text.buildSpannedString
 import com.godamy.myplayer.R
 import com.godamy.myplayer.domain.MediaItem
-import com.godamy.myplayer.framework.common.appendInfo
+import com.godamy.myplayer.ui.common.appendInfo
 
 class DetailInfoView @JvmOverloads constructor(
     context: Context,
